requirements.txt
pandas
polars
python-chess
joblib
tqdm
transformers
ms-swift
deepspeed